requirements.txt
beautifulsoup4==4.12.2
Pillow==9.5.0
